% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams2rqti-package.R
\docType{package}
\name{exams2rqti-package}
\alias{exams2rqti}
\alias{exams2rqti-package}
\title{Translate exams exercises to rqti objects}
\description{
An experimental adapter. Rendering belongs to exams and QTI
serialization belongs to rqti. Loading this package does not render exercises,
write files, or start a preview server.
}
\author{
\strong{Maintainer}: Johannes Titz \email{johannes.titz@gmail.com} (\href{https://orcid.org/0000-0002-1102-5719}{ORCID})

Authors:
\itemize{
  \item Johannes Titz \email{johannes.titz@gmail.com} (\href{https://orcid.org/0000-0002-1102-5719}{ORCID})
}

}
\keyword{internal}
