% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{ExamsComposite}
\alias{ExamsComposite}
\alias{ExamsComposite-class}
\alias{ExamsManual-class}
\alias{ExamsVerbatimNumeric-class}
\alias{createItemBody,ExamsComposite-method}
\alias{createResponseDeclaration,ExamsComposite-method}
\alias{createOutcomeDeclaration,ExamsComposite-method}
\alias{createResponseProcessing,ExamsComposite-method}
\alias{createResponseDeclaration,ExamsManual-method}
\alias{createItemBody,ExamsManual-method}
\alias{createResponseProcessing,ExamsVerbatimNumeric-method}
\alias{createOutcomeDeclaration,ExamsVerbatimNumeric-method}
\alias{createText,ExamsVerbatimNumeric-method}
\title{Composite exams interactions}
\usage{
\S4method{createItemBody}{ExamsComposite}(object)

\S4method{createResponseDeclaration}{ExamsComposite}(object)

\S4method{createOutcomeDeclaration}{ExamsComposite}(object)

\S4method{createResponseProcessing}{ExamsComposite}(object)

\S4method{createResponseDeclaration}{ExamsManual}(object)

\S4method{createItemBody}{ExamsManual}(object)

\S4method{createResponseProcessing}{ExamsVerbatimNumeric}(object)

\S4method{createOutcomeDeclaration}{ExamsVerbatimNumeric}(object)

\S4method{createText}{ExamsVerbatimNumeric}(object)
}
\arguments{
\item{object}{An adapter interaction or composite assessment item.}
}
\value{
The methods return QTI tags for rqti's serializer.
}
\description{
These rqti extensions combine numeric, text, choice, essay and upload
responses in one item. Construct them with \code{\link[=asRqtiItem]{asRqtiItem()}}. Manual scores
are deliberately left unset until supplied by the delivery/scoring system.
}
